requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
matplotlib==3.8.2
plotly==5.18.0
aiofiles==23.2.1
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
kaleido==0.2.1
reportlab==4.0.7
pillow==10.1.0
